+++ src/components/WidgetForm/SimpleSearch/filters/SimpleSearchUUIDFilterField.js
@@ -60,20 +60,10 @@ const SimpleSearchUUIDFilterField = ({
             <div className={css.flexContainer}>
               <div className={css.radioButton}>
                 <Field
+                  component={RadioButton}
                   defaultValue="absolute"
+                  disabled={isSetOrUnset}
                   name={`${name}.relativeOrAbsolute`}
-                  render={({ input }) => {
-                    return (
-                      <RadioButton
-                        checked={input.checked}
-                        disabled={isSetOrUnset}
-                        id="relative"
-                        name={input.name}
-                        onChange={input.onChange}
-                        value="relative"
-                      />
-                    );
-                  }}
                   type="radio"
                   validateFields={[`${name}.filterValue`]}
                   value="relative"
@@ -94,20 +84,9 @@ const SimpleSearchUUIDFilterField = ({
             <div className={css.flexContainer}>
               <div className={css.radioButton}>
                 <Field
+                  component={RadioButton}
+                  disabled={isSetOrUnset}
                   name={`${name}.relativeOrAbsolute`}
-                  render={({ input }) => {
-                    return (
-                      <RadioButton
-                        checked={input.checked}
-                        disabled={isSetOrUnset}
-                        id="absolute"
-                        name={input.name}
-                        onChange={input.onChange}
-                        validateFields={[`${name}.filterValue`]}
-                        value="absolute"
-                      />
-                    );
-                  }}
                   type="radio"
                   validateFields={[`${name}.filterValue`]}
                   value="absolute"